Fix more babel version cruft
.babelrc still referred to pre-babel v7 (and thus pre-named
scope packages, versus the named scope @babel packages), but
the .babelrc file itself is deprecated (by recommendation)
and was replaced by the spec compliant babel.config.js.

Package and lock files updated to reflect this. Notably, not
performing a complete `npm install` but merely installing the
new @babel packages, so as not to upgrade too much since that
tends to break Javascript things.

Fragility like crazy, but at least it's reproducible.